support arange
